% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch.R
\name{BirchCF}
\alias{BirchCF}
\title{Create a clustering feature tree using the BIRCH algorithm}
\usage{
BirchCF(x, Type = "df", branchingfactor = 4, threshold = 0.15)
}
\arguments{
\item{x}{Input to BirchCF function in dataframe format}

\item{Type}{default to 'df'}

\item{branchingfactor}{default to 4}

\item{threshold}{default to 0.15}
}
\value{
A hierarchical clustering feature tree
}
\description{
Function to create a clustering feature tree using the BIRCH algorithm
}
