% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSEM-package.R
\docType{package}
\name{ClusterSEM-package}
\alias{ClusterSEM}
\alias{ClusterSEM-package}
\title{ClusterSEM: Confirmatory factor analysis based on GWAS summary statistics}
\description{
Confirmatory factor analysis based on GWAS summary statistics.
}
\keyword{internal}
