% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcalg.R
\name{CLusterSEM.indepTest}
\alias{CLusterSEM.indepTest}
\title{Independence test for genetic covariance structure}
\usage{
CLusterSEM.indepTest(suffStat, x, y, S)
}
\arguments{
\item{suffStat}{Output from the multivariate LDSC/HDL/metaCCA function}

\item{x}{An item from 1:nrow(suffStat$S)}

\item{y}{An item from 1:nrow(suffStat$S)}

\item{S}{A set from 1:nrow(suffStat$S)}
}
\value{
p value indicating independence between x and y
}
\description{
Function to independence test for genetic covariance structure
}
