% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcalg.R
\name{ClusterSEM.suffStat}
\alias{ClusterSEM.suffStat}
\title{Re-sampling from the genetic covariance matrix and variance covariance matrix of the parameter estimates in S (i.e., LDSC/HDL/metaCCA)}
\usage{
ClusterSEM.suffStat(suffStat, seed = NULL, rep = 250)
}
\arguments{
\item{suffStat}{Output from the multivariate LDSC/HDL/metaCCA function}

\item{seed}{default to NULL}

\item{rep}{default to 250}
}
\value{
Re-sampled genetic covariance matrix, # of matrix = rep
}
\description{
Function for re-sampling from the genetic covariance matrix and variance covariance matrix of the parameter estimates in S
}
