% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_extract_instruments.R
\name{get_coef_from_gwas}
\alias{get_coef_from_gwas}
\title{Load data from IEU GWAS summary dataset}
\usage{
get_coef_from_gwas(exposure = "ieu-a-2", outcome = "ieu-a-7")
}
\arguments{
\item{exposure}{default to "ieu-a-2"}

\item{outcome}{default to "ieu-a-7"}
}
\value{
harmonised data
}
\description{
Estimate the causal effects of body mass index || id:ieu-a-2 on coronary artery disease || id:ieu-a-7 risk using 79 uncorrelated SNP genetic variants
}
