% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{ldsc}
\alias{ldsc}
\title{Estimate a genetic covariance structure using multivariate LD score regression}
\usage{
ldsc(
  traits,
  sample.prev,
  population.prev,
  ld,
  wld,
  trait.names = NULL,
  sep_weights = FALSE,
  chr = 22,
  n.blocks = 200,
  ldsc.log = NULL,
  stand = TRUE,
  select = FALSE,
  chisq.max = NA
)
}
\arguments{
\item{traits}{A vector of strings which point to LDSC munged files for trait you want to include in a SEM model}

\item{sample.prev}{A vector of sample prevalence for dichotomous traits and NA for continuous traits}

\item{population.prev}{A vector of population prevalence for dichotomous traits and NA for continuous traits}

\item{ld}{String which contains the path to the folder in which the LD scores used in the analysis are located.}

\item{wld}{String which contains the path to the folder in which the LD score weights used in the analysis are located}

\item{trait.names}{A character vector specifying how the traits should be named in the genetic covariance (S) matrix. These variable names can subsequently be used in later steps for model specification. If no value is provided, the function will automatically name the variables using the generic from of V1-VX.}

\item{sep_weights}{Logical which indicates whether the weights are different form the LD scores used for the regression, defaults to FALSE.}

\item{chr}{defaults to 22}

\item{n.blocks}{defaults to 200}

\item{ldsc.log}{defaults to NULL}

\item{stand}{If you want to output the standard errors (SE) of the ld-score regression in the order they are listed in the genetic covariance (i.e., S) matrix. Default to TRUE.}

\item{select}{default to FALSE}

\item{chisq.max}{default to NA}
}
\value{
The function returns a list with 5 named entries

S  estimated genetic covariance matrix

V  variance covariance matrix of the parameter estimates in S

I  matrix containing the "cross trait intercepts", or the error covariance between traits induced by overlap, in terms of subjects, between the GWAS on which the analyses are based

N  a vector containing the sample size (for the genetic variances) and the geometric mean of sample sizes (i.e. sqrt(N1,N2)) between two samples for the covariance

m  number of SNPs used to compute the LD scores
}
\description{
Function to run LD score regression (https://github.com/bulik/ldsc) to compute the genetic covariance between a series of traits based on genome wide summary statistics obtained from GWAS.
}
