% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_parallel.R
\name{ldsc_parallel}
\alias{ldsc_parallel}
\title{Parallelly estimate a genetic covariance structure using multivariate LD score regression}
\usage{
ldsc_parallel(
  traits,
  sample.prev,
  population.prev,
  ld,
  wld,
  trait.names = NULL,
  sep_weights = FALSE,
  chr = 22,
  n.blocks = 200,
  ldsc.log = NULL,
  stand = TRUE,
  select = FALSE,
  chisq.max = NA,
  save.output = FALSE,
  output.path = NULL,
  parallel = TRUE,
  cores = NULL
)
}
\arguments{
\item{traits}{A vector of strings which point to LDSC munged files for trait you want to include in a SEM model}

\item{sample.prev}{A vector of sample prevalence for dichotomous traits and NA for continuous traits}

\item{population.prev}{A vector of population prevalence for dichotomous traits and NA for continuous traits}

\item{ld}{String which contains the path to the folder in which the LD scores used in the analysis are located.}

\item{wld}{String which contains the path to the folder in which the LD score weights used in the analysis are located}

\item{trait.names}{A character vector specifying how the traits should be named in the genetic covariance matrix (i.e., S). These variable names can subsequently be used in later steps for model specification. If no value is provided, the function will automatically name the variables using the generic from of V1-VX.}

\item{sep_weights}{Logical which indicates whether the weights are different form the LD scores used for the regression, defaults to FALSE.}

\item{chr}{defaults to 22}

\item{n.blocks}{defaults to 200}

\item{ldsc.log}{What to name the .log file if you want to override. Default to name the .log file based on file names used as input.}

\item{stand}{If you want to output the standard errors (SE) of the ld-score regression in the order they are listed in the genetic covariance matrix (i.e., S). Default to TRUE.}

\item{select}{default to FALSE, option: EVEN, ODD, c(...)}

\item{chisq.max}{default to NA}

\item{save.output}{default to FALSE, which indicates whether the output of ldsc function is saved to a .RData file named as ldsc.log.}

\item{output.path}{default to NULL (the current working directory). If you want to specify other output.path, please make output.path string ending with '/'.}

\item{parallel}{default to FALSE}

\item{cores}{default to 1}
}
\value{
The function returns a list with 5 named entries

S  estimated genetic variance/covariance matrix

SE variance matrix of the estimated parameter in S

I  matrix containing the "cross trait intercepts", or the error covariance between traits induced by overlap, in terms of subjects, between the GWAS on which the analyses are based

N  a vector containing the sample size for the genetic variances (i.e., heritability) and the geometric mean of sample sizes (i.e., sqrt(N1*N2)) between two samples for the genetic covariance

m  number of SNPs used to compute the LD scores
}
\description{
Function to run LD score regression (https://github.com/bulik/ldsc) to compute the genetic covariance between a series of traits based on genome wide summary statistics obtained from GWAS.
}
\references{
Consortium, S. W. G. of the P. G. et al. LD Score regression distinguishes confounding from polygenicity in genome-wide association studies. Nat. Genet. 47, 291–295 (2015).
}
