% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrSEM-package.R
\docType{package}
\name{mrSEM-package}
\alias{mrSEM}
\alias{mrSEM-package}
\title{mrSEM: Confirmatory factor analysis based on GWAS summary statistics}
\description{
Confirmatory factor analysis based on GWAS summary statistics.
}
\keyword{internal}
