% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{munge}
\alias{munge}
\title{Clean and munge files to enable LD score regression}
\usage{
munge(
  files,
  hm3,
  trait.names = NULL,
  N = NULL,
  info.filter = 0.9,
  maf.filter = 0.01,
  log.name = NULL,
  column.names = list(),
  parallel = FALSE,
  cores = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{files}{A list of file names, files must be located in the working directory, or a path must be provided.}

\item{hm3}{A file (UNZIPPED) of HAPMAP3 SNPs with some basic cleaning applied it is supplied and created by the original LD score regression developers and available here: https://data.broadinstitute.org/alkesgroup/LDSCORE/w_hm3.snplist.bz2}

\item{trait.names}{A vector of trait names which will be used as names for the munged files}

\item{N}{default to NULL}

\item{info.filter}{Numeric value which is used as a lower bound for imputation quality (INFO)}

\item{maf.filter}{Numeric value used as a lower bound for minor allele frequency}

\item{log.name}{default to NULL}

\item{column.names}{column.names is a list(SNP, A1, A2, effect, P, MAF, Z) to specify other name}

\item{parallel}{default to FALSE}

\item{cores}{default to NULL}

\item{overwrite}{default to TRUE}
}
\value{
The function writes files of the ".sumstats.gz" format, which can be used to estimate SNP heritability and genetic covariance using the ldsc() function. The function will also output a .log file that should be examined to ensure that column names are being interpret correctly.
}
\description{
Function to process GWAS summary statistics files and prepare them for LD score regression
}
