% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_extract_instruments.R
\name{mv_get_coef_from_gwas}
\alias{mv_get_coef_from_gwas}
\title{Load multi data from IEU GWAS summary dataset}
\usage{
mv_get_coef_from_gwas(
  exposures = c("ieu-a-299", "ieu-a-300", "ieu-a-302"),
  outcome = "ieu-a-7"
)
}
\arguments{
\item{exposures}{default to c("ieu-a-299","ieu-a-300","ieu-a-302")}

\item{outcome}{default to "ieu-a-7"}
}
\value{
harmonised data
}
\description{
Estimate the multivariate effects of HDL cholesterol || id:ieu-a-299 (#SNP=79), LDL cholesterol || id:ieu-a-300 (#SNP=68) and triglycerides || id:ieu-a-302 (#SNP=42) on coronary artery disease || id:ieu-a-7 risk using 143 uncorrelated SNP genetic variants
}
