% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_ldsc.R
\name{s_ldsc}
\alias{s_ldsc}
\title{Estimate genetic covariance structure within functional annotations using multivariate Stratified LD score regression}
\usage{
s_ldsc(
  traits,
  sample.prev = NULL,
  population.prev = NULL,
  ld,
  wld,
  frq,
  trait.names = NULL,
  n.blocks = 200,
  ldsc.log = NULL,
  exclude_cont = TRUE
)
}
\arguments{
\item{traits}{A vector of file names which point to  LDSC munged files for trait you want to include}

\item{sample.prev}{A vector of sample prevalence for dichotomous traits and \code{NA} for continuous traits. Default = NULL.}

\item{population.prev}{A vector of population prevalence for dichotomous traits and \code{NA} for continuous traits. Default = NULL.}

\item{ld}{A folder (or folders) of partitioned LD scores used as the independent variable in S-LDSC.}

\item{wld}{A folder of non-partitioned LD scores used as regression weights.}

\item{frq}{A folder of allele frequency files.}

\item{trait.names}{A character vector specifying how the traits should be named. These variable names can subsequently be used in later steps for model specification.}

\item{n.blocks}{Number of blocks to use for the jacknive procedure which is used to estiamte entries in V, higher values will be optimal if you have a large number of variables and also slower, defaults to 200.}

\item{ldsc.log}{What to name the .log file if you want to override. Default to name file based on file names used as input.}

\item{exclude_cont}{Whether to exclude continuous annotations from S-LDSC estimation.}
}
\value{
The function returns a list with 9 named entries

S  The zero-order genetic covariance matrices for each annotation

V  The zero-order sampling covariance matrices for each annotation

S_Tau  The tau matrices for each annotation

V_Tau  The tau sampling covariance matrices for each annotation

I  matrix containing the "cross trait intercepts", or the error covariance between traits induced by overlap, in terms of subjects, between the GWAS on which the analyses are based

N  a vector containing the sample size (for the genetic variances) and the geometric mean of sample sizes (i.e. sqrt(N1,N2)) between two samples for the covariance
}
\description{
Function to run Stratified LD score regression
}
