% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usermodel.R
\name{usermodel}
\alias{usermodel}
\title{Run user-specified structural equation model on genetic covariance matrix}
\usage{
usermodel(
  covstruc,
  estimation = "DWLS",
  model = "",
  CFIcalc = TRUE,
  std.lv = FALSE,
  imp_cov = FALSE,
  fix_resid = TRUE,
  toler = NULL
)
}
\arguments{
\item{covstruc}{genetic covariance matrix}

\item{estimation}{Options are either Diagonally Weighted Least Squares ("DWLS"; the default) or Maximum Likelihood ("ML")}

\item{model}{Model to be specified using lavaan syntax}

\item{CFIcalc}{Optional argument to denote whether CFI is requested (default = TRUE). In some cases the estimation of the independent (i.e., Null) model for calculation of CFI can be time consuming. If the function seems to be stuck on this step, we would suggest re-running with this option set to FALSE}

\item{std.lv}{Optional argument to denote whether all latent variables are standardized using unit variance identification (default = FALSE)}

\item{imp_cov}{Optional argument to denote whether the user wants the model-implied and residual covariance matrix included in the usermodel output (default = FALSE)}

\item{fix_resid}{default to TRUE}

\item{toler}{default to NULL}
}
\value{
The function estimates a user-specified model, along with model fit indices.
}
\description{
Function to run a user-specified structural equation model based on output from multivariable LDSC, HDL and metaCCA
}
