% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.model.R
\name{write.model}
\alias{write.model}
\title{Automate writing lavaan model syntax based on clustering output}
\usage{
write.model(
  clusters,
  S_LD,
  fix_resid = TRUE,
  common = FALSE,
  hierarchical = FALSE
)
}
\arguments{
\item{clusters}{Output from various clustering algorithms}

\item{S_LD}{Genetic covariance matrix}

\item{fix_resid}{default to TRUE}

\item{common}{Whether to specify a common factor model, default to FALSE}

\item{hierarchical}{Whether to specify a hierarchical exploratory factor model, default to FALSE}
}
\value{
lavaan model syntax satisfying clusters
}
\description{
Function to automate writing lavaan model syntax based on clustering output
}
