% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcalg.R
\name{ClusterSEM.indepTest}
\alias{ClusterSEM.indepTest}
\title{Independence test for genetic covariance structure}
\usage{
ClusterSEM.indepTest(x, y, S, suffStat)
}
\arguments{
\item{x}{An item from 1:nrow(suffStat$S)}

\item{y}{An item from 1:nrow(suffStat$S)}

\item{S}{A set from 1:nrow(suffStat$S)}

\item{suffStat}{Output from the multivariate LDSC/HDL/metaCCA function}
}
\value{
p value indicating independence between x and y
}
\description{
Function to independence test for genetic covariance structure
}
